// frontend/src/js/shared/mavlink2rest-ts/messages/mavlink2rest-enum.ts
export enum MavAutopilot {
  // Generic autopilot, full support for everything
  MAV_AUTOPILOT_GENERIC = 'MAV_AUTOPILOT_GENERIC',
  // Reserved for future use.
  MAV_AUTOPILOT_RESERVED = 'MAV_AUTOPILOT_RESERVED',
  // SLUGS autopilot, http://slugsuav.soe.ucsc.edu
  MAV_AUTOPILOT_SLUGS = 'MAV_AUTOPILOT_SLUGS',
  // ArduPilot - Plane/Copter/Rover/Sub/Tracker, https://ardupilot.org
  MAV_AUTOPILOT_ARDUPILOTMEGA = 'MAV_AUTOPILOT_ARDUPILOTMEGA',
  // OpenPilot, http://openpilot.org
  MAV_AUTOPILOT_OPENPILOT = 'MAV_AUTOPILOT_OPENPILOT',
  // Generic autopilot only supporting simple waypoints
  MAV_AUTOPILOT_GENERIC_WAYPOINTS_ONLY = 'MAV_AUTOPILOT_GENERIC_WAYPOINTS_ONLY',
  // Generic autopilot supporting waypoints and other simple navigation commands
  MAV_AUTOPILOT_GENERIC_WAYPOINTS_AND_SIMPLE_NAVIGATION_ONLY = 'MAV_AUTOPILOT_GENERIC_WAYPOINTS_AND_SIMPLE_NAVIGATION_ONLY',
  // Generic autopilot supporting the full mission command set
  MAV_AUTOPILOT_GENERIC_MISSION_FULL = 'MAV_AUTOPILOT_GENERIC_MISSION_FULL',
  // No valid autopilot, e.g. a GCS or other MAVLink component
  MAV_AUTOPILOT_INVALID = 'MAV_AUTOPILOT_INVALID',
  // PPZ UAV - http://nongnu.org/paparazzi
  MAV_AUTOPILOT_PPZ = 'MAV_AUTOPILOT_PPZ',
  // UAV Dev Board
  MAV_AUTOPILOT_UDB = 'MAV_AUTOPILOT_UDB',
  // FlexiPilot
  MAV_AUTOPILOT_FP = 'MAV_AUTOPILOT_FP',
  // PX4 Autopilot - http://px4.io/
  MAV_AUTOPILOT_PX4 = 'MAV_AUTOPILOT_PX4',
  // SMACCMPilot - http://smaccmpilot.org
  MAV_AUTOPILOT_SMACCMPILOT = 'MAV_AUTOPILOT_SMACCMPILOT',
  // AutoQuad -- http://autoquad.org
  MAV_AUTOPILOT_AUTOQUAD = 'MAV_AUTOPILOT_AUTOQUAD',
  // Armazila -- http://armazila.com
  MAV_AUTOPILOT_ARMAZILA = 'MAV_AUTOPILOT_ARMAZILA',
  // Aerob -- http://aerob.ru
  MAV_AUTOPILOT_AEROB = 'MAV_AUTOPILOT_AEROB',
  // ASLUAV autopilot -- http://www.asl.ethz.ch
  MAV_AUTOPILOT_ASLUAV = 'MAV_AUTOPILOT_ASLUAV',
  // SmartAP Autopilot - http://sky-drones.com
  MAV_AUTOPILOT_SMARTAP = 'MAV_AUTOPILOT_SMARTAP',
  // AirRails - http://uaventure.com
  MAV_AUTOPILOT_AIRRAILS = 'MAV_AUTOPILOT_AIRRAILS',
}

/**
 * MAVLINK component type reported in HEARTBEAT message.
 * Flight controllers must report the type of the vehicle on which they are mounted (e.g. MAV_TYPE_OCTOROTOR).
 * All other components must report a value appropriate for their type (e.g. a camera must use MAV_TYPE_CAMERA).
 */
export enum MavType {
  // Generic micro air vehicle
  MAV_TYPE_GENERIC = 'MAV_TYPE_GENERIC',
  // Fixed wing aircraft.
  MAV_TYPE_FIXED_WING = 'MAV_TYPE_FIXED_WING',
  // Quadrotor
  MAV_TYPE_QUADROTOR = 'MAV_TYPE_QUADROTOR',
  // Coaxial helicopter
  MAV_TYPE_COAXIAL = 'MAV_TYPE_COAXIAL',
  // Normal helicopter with tail rotor.
  MAV_TYPE_HELICOPTER = 'MAV_TYPE_HELICOPTER',
  // Ground installation
  MAV_TYPE_ANTENNA_TRACKER = 'MAV_TYPE_ANTENNA_TRACKER',
  // Operator control unit / ground control station
  MAV_TYPE_GCS = 'MAV_TYPE_GCS',
  // Airship, controlled
  MAV_TYPE_AIRSHIP = 'MAV_TYPE_AIRSHIP',
  // Free balloon, uncontrolled
  MAV_TYPE_FREE_BALLOON = 'MAV_TYPE_FREE_BALLOON',
  // Rocket
  MAV_TYPE_ROCKET = 'MAV_TYPE_ROCKET',
  // Ground rover
  MAV_TYPE_GROUND_ROVER = 'MAV_TYPE_GROUND_ROVER',
  // Surface vessel, boat, ship
  MAV_TYPE_SURFACE_BOAT = 'MAV_TYPE_SURFACE_BOAT',
  // Submarine
  MAV_TYPE_SUBMARINE = 'MAV_TYPE_SUBMARINE',
  // Hexarotor
  MAV_TYPE_HEXAROTOR = 'MAV_TYPE_HEXAROTOR',
  // Octorotor
  MAV_TYPE_OCTOROTOR = 'MAV_TYPE_OCTOROTOR',
  // Tricopter
  MAV_TYPE_TRICOPTER = 'MAV_TYPE_TRICOPTER',
  // Flapping wing
  MAV_TYPE_FLAPPING_WING = 'MAV_TYPE_FLAPPING_WING',
  // Kite
  MAV_TYPE_KITE = 'MAV_TYPE_KITE',
  // Onboard companion controller
  MAV_TYPE_ONBOARD_CONTROLLER = 'MAV_TYPE_ONBOARD_CONTROLLER',
  // Two-rotor VTOL using control surfaces in vertical operation in addition. Tailsitter.
  MAV_TYPE_VTOL_DUOROTOR = 'MAV_TYPE_VTOL_DUOROTOR',
  // Quad-rotor VTOL using a V-shaped quad config in vertical operation. Tailsitter.
  MAV_TYPE_VTOL_QUADROTOR = 'MAV_TYPE_VTOL_QUADROTOR',
  // Tiltrotor VTOL
  MAV_TYPE_VTOL_TILTROTOR = 'MAV_TYPE_VTOL_TILTROTOR',
  // VTOL reserved 2
  MAV_TYPE_VTOL_RESERVED2 = 'MAV_TYPE_VTOL_RESERVED2',
  // VTOL reserved 3
  MAV_TYPE_VTOL_RESERVED3 = 'MAV_TYPE_VTOL_RESERVED3',
  // VTOL reserved 4
  MAV_TYPE_VTOL_RESERVED4 = 'MAV_TYPE_VTOL_RESERVED4',
  // VTOL reserved 5
  MAV_TYPE_VTOL_RESERVED5 = 'MAV_TYPE_VTOL_RESERVED5',
  // Gimbal
  MAV_TYPE_GIMBAL = 'MAV_TYPE_GIMBAL',
  // ADSB system
  MAV_TYPE_ADSB = 'MAV_TYPE_ADSB',
  // Steerable, nonrigid airfoil
  MAV_TYPE_PARAFOIL = 'MAV_TYPE_PARAFOIL',
  // Dodecarotor
  MAV_TYPE_DODECAROTOR = 'MAV_TYPE_DODECAROTOR',
  // Camera
  MAV_TYPE_CAMERA = 'MAV_TYPE_CAMERA',
  // Charging station
  MAV_TYPE_CHARGING_STATION = 'MAV_TYPE_CHARGING_STATION',
  // FLARM collision avoidance system
  MAV_TYPE_FLARM = 'MAV_TYPE_FLARM',
  // Servo
  MAV_TYPE_SERVO = 'MAV_TYPE_SERVO',
  // Open Drone ID. See https://mavlink.io/en/services/opendroneid.html.
  MAV_TYPE_ODID = 'MAV_TYPE_ODID',
}
// ...
export enum MavModeFlag {
  //  MAV safety set to armed. Motors are enabled / running / can start. Ready to fly. Additional note: this flag is to be ignore when sent in the command MAV_CMD_DO_SET_MODE and MAV_CMD_COMPONENT_ARM_DISARM shall be used instead. The flag can still be used to report the armed state.
  MAV_MODE_FLAG_SAFETY_ARMED = 0b10000000,
  //  Remote control input is enabled.
  MAV_MODE_FLAG_MANUAL_INPUT_ENABLED = 0b01000000,
  //  Hardware in the loop simulation. All motors / actuators are blocked, but internal software is full operational.
  MAV_MODE_FLAG_HIL_ENABLED = 0b00100000,
  //  System stabilizes electronically its attitude (and optionally position). It needs however further control inputs to move around.
  MAV_MODE_FLAG_STABILIZE_ENABLED = 0b00010000,
  //  Guided mode enabled, system flies waypoints / mission items.
  MAV_MODE_FLAG_GUIDED_ENABLED = 0b00001000,
  //  Autonomous mode enabled, system finds its own goal positions. Guided flag can be set or not, depends on the actual implementation.
  MAV_MODE_FLAG_AUTO_ENABLED = 0b00000100,
  //  System has a test mode enabled. This flag is intended for temporary system tests and should not be used for stable implementations.
  MAV_MODE_FLAG_TEST_ENABLED = 0b00000010,
  //  Reserved for future use.
  MAV_MODE_FLAG_CUSTOM_MODE_ENABLED = 0b00000001,
}
// ...
export enum MavState {
  // Uninitialized system, state is unknown.
  MAV_STATE_UNINIT = 'MAV_STATE_UNINIT',
  // System is booting up.
  MAV_STATE_BOOT = 'MAV_STATE_BOOT',
  // System is calibrating and not flight-ready.
  MAV_STATE_CALIBRATING = 'MAV_STATE_CALIBRATING',
  // System is grounded and on standby. It can be launched any time.
  MAV_STATE_STANDBY = 'MAV_STATE_STANDBY',
  // System is active and might be already airborne. Motors are engaged.
  MAV_STATE_ACTIVE = 'MAV_STATE_ACTIVE',
  // System is in a non-normal flight mode. It can however still navigate.
  MAV_STATE_CRITICAL = 'MAV_STATE_CRITICAL',
  // System is in a non-normal flight mode. It lost control over parts or over the whole airframe. It is in mayday and going down.
  MAV_STATE_EMERGENCY = 'MAV_STATE_EMERGENCY',
  // System just initialized its power-down sequence, will shut down now.
  MAV_STATE_POWEROFF = 'MAV_STATE_POWEROFF',
  // System is terminating itself.
  MAV_STATE_FLIGHT_TERMINATION = 'MAV_STATE_FLIGHT_TERMINATION',
}